requirements.txt
websockets
websocket_client
requests
NorenRestApi
python-dotenv